+++ packages/expo-router/CHANGELOG.md
@@ -15,6 +15,7 @@
 ### 🎉 New features
 
 - [ios] add support for renderingMode of xcassets in header items and bottom tabs ([#43185](https://github.com/expo/expo/pull/43185) by [@Ubax](https://github.com/Ubax))
+- use scoped dynamic options in stack ([#43184](https://github.com/expo/expo/pull/43184) by [@Ubax](https://github.com/Ubax))
 
 ### 🐛 Bug fixes